Move to browserify from brunch. Add Hound
Move away from brunch to browserify, this is
basically so that I can have a nicer environment
for running the Vue.js code.

Note also the start_browserify_workers script
which is used to start up the npm scripts and then
shut them down correctly when the elixir process stops.

This is done by looking for when the stdin is closed
and then killing off our child processes. It's not
necessarily bombproof, but it does seem to work ok.

Hound is added for feature tests.