Move to download files in a stream.
Now we download the files using stream_to:
and then enter a receive loop.

Seems to work nicely enough, though there is
apparently no way to use "back pressure", to
tell the stream to back off. Still, with the
downloading I'm not even sure how you'd do that
in the first place, I mean the server is just
going to push the file down to you as fast
as it can.

Whether I should (and I should) be using a
job queue instead - or at least in some way
limiting the number of downloads, well, that's
something for the future. Not too hard I think
now that I have the basics figured out.