Fix for youtube-dl sometimes altering filename
Sometimes youtube-dl will alter a filename when it
attempts to download it and it turns out to require
a different format.

It wasn't super straightforward to me how to get back
the rewritten filename in a standard way. So, I cheat.

I now prepend the unique-id to the start of the filename,
and then, even if the extension is altered as a part of the
downloading I can find it.

So, after the file has been downloaded, I do a wildcard
match on all the filenames starting with the specific ID
and then I assume that there's only one match and that's the
one I want.

I will alter this code to raise an exception if there's more
than one. I think.